remove database package [builds]
